app/requirements/base.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.27
pydantic==2.11.7
pydantic-settings==2.2.1

# AWS & Cloud Services
boto3==1.34.106
pyathena==3.0.10

# Async & Performance
uvloop==0.21.0
httptools==0.6.4
nest-asyncio==1.6.0

# Data Processing
pandas==2.1.4
pymongo==4.6.1
motor==3.3.2

# Redis
redis[hiredis]==5.0.1

# HTTP & API
requests==2.29.0

# CLI
typer==0.15.1

# Async Django
adrf==0.1.8

# Monitoring & Observability
elastic-apm==6.18.0
datadog==0.50.2

# Utilities
pytz==2024.1
python-decouple==3.8
base58==2.1.1
